+++ src/framework.js
@@ -38,6 +38,7 @@ export { optimizeImage, batchDOMOperations } from './utils/dom.js';
 // First-party plugins
 export { iconsPlugin } from './plugins/icons/icons-plugin.js';
 export { themePlugin } from './plugins/theme/theme-plugin.js';
+export { alertsPlugin } from './plugins/alerts/alerts-plugin.js';
 
 // Import classes for internal use
 import { ComponentManager } from './core/component-manager.js';
@@ -313,5 +314,5 @@ export function createApp(config = {}) {
 }
 
 // Framework metadata
-export const FRAMEWORK_VERSION = '1.2.0';
+export const FRAMEWORK_VERSION = '1.3.0';
 export const FRAMEWORK_NAME = 'VanillaForge';